use chapoly suites only when client prefers
The upside of this over the equal preference group patch is it's a
much simpler patch to vet and trust.

The downside is it's hacky and not very general, and doesn't allow
the server ciphersuite config to intentionally prioritize chapoly
over aes-gcm regardless of the client's prefs.  At the very least,
this patch would be better if the behavior were controlled with an
SSL_OP_FOO.

This is based on how cloudflare does things in their 1.0.2 patch:
https://github.com/cloudflare/sslconfig/blob/master/patches/openssl__chacha20_poly1305_draft_and_rfc_ossl102g.patch#L4929

Ref: openssl/openssl#541