increase DEFAULT_BUFFER_SIZE from 4K to 8K
The historical SSLeay value was 1K.  This was increased to 4K in
for release 0.9.6b back in 2001.  Today, it is not uncommon for
server handshake responses to exceed 4K given intermediate chain
certs and OCSP stapling, but typical sizes are still within 8K.

When the server response exceeds the buffer size, the TLS
handshake suffers an extra RTT of latency penalty.  It is also
difficult (and shouldn't be necessary given how common this is) to
correct this manually using the normal APIs for a server
application.

There may have also been some rationale around TCP IW3 limits in
setting the previous 4K size, but modern servers are typically
IW10, which allows the server to initially send 8K easily.  IW10
allows for at most 14600 in the ideal case, but I wouldn't expect
it to commonly be any less than 12200 (IPv6 minimum MTU case).